+++ src/main/java/com/gdsc/knu/repository/AchievementRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AchievementRepository extends JpaRepository<Achievement, Integer> {
-    Optional<Achievement> findByUserId(Integer userID);
+public interface AchievementRepository extends JpaRepository<Achievement, Long> {
+    Optional<Achievement> findByUserId(Long userID);
 }
